Step 64.
Create a for loop in robot.Sense() that iterates over all the SENSOR instances stored in the dictionary self.sensors.

My code threw a type error:
    def Sense(self,timeStep):
        for i in self.sensors:
            self.sensors[i] = SENSOR.Get_Value(timeStep)

        pass

Correted code:
The issue lies in the Sense method of your ROBOT class. Specifically, you are trying to call SENSOR.Get_Value(timeStep) as if it were a static method, but it is an instance method. You should call Get_Value on each instance of SENSOR stored in self.sensors.

def Sense(self, timeStep):
    for sensor in self.sensors.values():
        sensor.Get_Value(timeStep)

https://www.reddit.com/r/ludobots/wiki/refactoring/